requirements.txt
Django>=2.2.6,<2.3.0
requests>=2.22.0,<2.23.0
beautifulsoup4>=4.8.1,<4.9.0
djangorestframework>=3.10.3,<3.11.0